refactor: Toast 관련 시간 상수 분리 #126
- 코드의 가독성을 높이기 위해 매직 넘버를 상수로 변경하여 파일을
  분리하였습니다
