+++ frontend/src/components/Toast/toastType.ts
